Extend use of global_evaluate
